// src/features/personas/models/Persona.ts
export interface Persona {
  Id: string;
  Rol: string;
  Email: string;
  Nombre: string;
  IdPais: string;
  NumDoc: string;
  Estado: string;
  TipoDoc: string;
  IdCiudad: string;
  Apellido: string;
  Telefono: string;
  Direccion: string;
  CuentaUsuario: string;
  FechaNacimiento: Date;
  IdDepartamento: string;
}